complete web3 stake buckets apis
