+++ action/stake_bucketsbyindexes.go
@@ -1,0 +1,145 @@
+package action
+
+import (
+	"strings"
+
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/iotexproject/iotex-proto/golang/iotexapi"
+	"github.com/iotexproject/iotex-proto/golang/iotextypes"
+	"google.golang.org/protobuf/proto"
+)
+
+const _bucketsByIndexesInterfaceABI = `[
+	{
+		"inputs": [
+			{
+				"internalType": "uint64[]",
+				"name": "indexes",
+				"type": "uint64[]"
+			}
+		],
+		"name": "bucketsByIndexes",
+		"outputs": [
+			{
+				"components": [
+					{
+						"internalType": "uint64",
+						"name": "index",
+						"type": "uint64"
+					},
+					{
+						"internalType": "address",
+						"name": "candidateAddress",
+						"type": "address"
+					},
+					{
+						"internalType": "uint256",
+						"name": "stakedAmount",
+						"type": "uint256"
+					},
+					{
+						"internalType": "uint32",
+						"name": "stakedDuration",
+						"type": "uint32"
+					},
+					{
+						"internalType": "int64",
+						"name": "createTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "int64",
+						"name": "stakeStartTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "int64",
+						"name": "unstakeStartTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "bool",
+						"name": "autoStake",
+						"type": "bool"
+					},
+					{
+						"internalType": "address",
+						"name": "owner",
+						"type": "address"
+					}
+				],
+				"internalType": "struct IStaking.VoteBucket[]",
+				"name": "",
+				"type": "tuple[]"
+			}
+		],
+		"stateMutability": "view",
+		"type": "function"
+	}
+]`
+
+var _bucketsByIndexesMethod abi.Method
+
+func init() {
+	_interface, err := abi.JSON(strings.NewReader(_bucketsByIndexesInterfaceABI))
+	if err != nil {
+		panic(err)
+	}
+	var ok bool
+	_bucketsByIndexesMethod, ok = _interface.Methods["bucketsByIndexes"]
+	if !ok {
+		panic("fail to load the method")
+	}
+}
+
+type BucketsByIndexesStateContext struct {
+	*baseStateContext
+}
+
+func newBucketsByIndexesStateContext(data []byte) (*BucketsByIndexesStateContext, error) {
+	paramsMap := map[string]interface{}{}
+	ok := false
+	if err := _bucketsMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
+		return nil, err
+	}
+	var index []uint64
+	if index, ok = paramsMap["indexes"].([]uint64); !ok {
+		return nil, ErrDecodeFailure
+	}
+
+	method := &iotexapi.ReadStakingDataMethod{
+		Method: iotexapi.ReadStakingDataMethod_BUCKETS_BY_INDEXES,
+	}
+	methodBytes, err := proto.Marshal(method)
+	if err != nil {
+		return nil, err
+	}
+	arguments := &iotexapi.ReadStakingDataRequest{
+		Request: &iotexapi.ReadStakingDataRequest_BucketsByIndexes{
+			BucketsByIndexes: &iotexapi.ReadStakingDataRequest_VoteBucketsByIndexes{
+				Index: index,
+			},
+		},
+	}
+	argumentsBytes, err := proto.Marshal(arguments)
+	if err != nil {
+		return nil, err
+	}
+	return &BucketsByIndexesStateContext{
+		&baseStateContext{
+			&Parameters{
+				MethodName: methodBytes,
+				Arguments:  [][]byte{argumentsBytes},
+			},
+		},
+	}, nil
+}
+
+func (r *BucketsByIndexesStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
+	var result iotextypes.VoteBucketList
+	if err := proto.Unmarshal(resp.Data, &result); err != nil {
+		return "", err
+	}
+
+	return encodeVoteBucketListToEth(result)
+}
