+++ action/stake_bucketsbyvoter.go
@@ -1,0 +1,172 @@
+package action
+
+import (
+	"strings"
+
+	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/iotexproject/iotex-address/address"
+	"github.com/iotexproject/iotex-proto/golang/iotexapi"
+	"github.com/iotexproject/iotex-proto/golang/iotextypes"
+	"google.golang.org/protobuf/proto"
+)
+
+var _bucketsByVoterInterfaceABI = `[
+	{
+		"inputs": [
+			{
+				"internalType": "address",
+				"name": "voter",
+				"type": "address"
+			},
+			{
+				"internalType": "uint32",
+				"name": "offset",
+				"type": "uint32"
+			},
+			{
+				"internalType": "uint32",
+				"name": "limit",
+				"type": "uint32"
+			}
+		],
+		"name": "bucketsByVoter",
+		"outputs": [
+			{
+				"components": [
+					{
+						"internalType": "uint64",
+						"name": "index",
+						"type": "uint64"
+					},
+					{
+						"internalType": "address",
+						"name": "candidateAddress",
+						"type": "address"
+					},
+					{
+						"internalType": "uint256",
+						"name": "stakedAmount",
+						"type": "uint256"
+					},
+					{
+						"internalType": "uint32",
+						"name": "stakedDuration",
+						"type": "uint32"
+					},
+					{
+						"internalType": "int64",
+						"name": "createTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "int64",
+						"name": "stakeStartTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "int64",
+						"name": "unstakeStartTime",
+						"type": "int64"
+					},
+					{
+						"internalType": "bool",
+						"name": "autoStake",
+						"type": "bool"
+					},
+					{
+						"internalType": "address",
+						"name": "owner",
+						"type": "address"
+					}
+				],
+				"internalType": "struct IStaking.VoteBucket[]",
+				"name": "",
+				"type": "tuple[]"
+			}
+		],
+		"stateMutability": "view",
+		"type": "function"
+	}
+]`
+
+var _bucketsByVoterMethod abi.Method
+
+func init() {
+	_interface, err := abi.JSON(strings.NewReader(_bucketsByVoterInterfaceABI))
+	if err != nil {
+		panic(err)
+	}
+	var ok bool
+	_bucketsByVoterMethod, ok = _interface.Methods["bucketsByVoter"]
+	if !ok {
+		panic("fail to load the method")
+	}
+}
+
+type BucketsByVoterStateContext struct {
+	*baseStateContext
+}
+
+func newBucketsByVoterStateContext(data []byte) (*BucketsByVoterStateContext, error) {
+	paramsMap := map[string]interface{}{}
+	ok := false
+	if err := _bucketsMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
+		return nil, err
+	}
+	var voter common.Address
+	if voter, ok = paramsMap["voter"].(common.Address); !ok {
+		return nil, ErrDecodeFailure
+	}
+	voterAddress, err := address.FromBytes(voter[:])
+	if err != nil {
+		return nil, err
+	}
+	var offset, limit uint32
+	if offset, ok = paramsMap["offset"].(uint32); !ok {
+		return nil, ErrDecodeFailure
+	}
+	if limit, ok = paramsMap["limit"].(uint32); !ok {
+		return nil, ErrDecodeFailure
+	}
+
+	method := &iotexapi.ReadStakingDataMethod{
+		Method: iotexapi.ReadStakingDataMethod_BUCKETS_BY_VOTER,
+	}
+	methodBytes, err := proto.Marshal(method)
+	if err != nil {
+		return nil, err
+	}
+	arguments := &iotexapi.ReadStakingDataRequest{
+		Request: &iotexapi.ReadStakingDataRequest_BucketsByVoter{
+			BucketsByVoter: &iotexapi.ReadStakingDataRequest_VoteBucketsByVoter{
+				VoterAddress: voterAddress.String(),
+				Pagination: &iotexapi.PaginationParam{
+					Offset: offset,
+					Limit:  limit,
+				},
+			},
+		},
+	}
+	argumentsBytes, err := proto.Marshal(arguments)
+	if err != nil {
+		return nil, err
+	}
+	return &BucketsByVoterStateContext{
+		&baseStateContext{
+			&Parameters{
+				MethodName: methodBytes,
+				Arguments:  [][]byte{argumentsBytes},
+			},
+		},
+	}, nil
+}
+
+func (r *BucketsByVoterStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
+	var result iotextypes.VoteBucketList
+	if err := proto.Unmarshal(resp.Data, &result); err != nil {
+		return "", err
+	}
+
+	return encodeVoteBucketListToEth(result)
+}
